fix/honestflix: resolve conflicts of backend restructuring
