+++ main.py
@@ -1,0 +1,10 @@
+#!/usr/bin/env python3
+"""
+Main entry point for the API Gateway
+"""
+
+from src.main import app
+
+if __name__ == "__main__":
+    import uvicorn
+    uvicorn.run(app, host="0.0.0.0", port=8000)
